requirements.txt
numpy==1.24
llvmlite
numba
torch>=1.1
tensorboardX
easydict
pyyaml
scikit-image
tqdm
torchvision
SharedArray
prettytable
